fix(preview): report connection failures
